Simplify sending of game events realtime. Rails commit: df3249a
Remove code associated with old process.
Now we simply use a unique token that is associated with socket.id only once
when game is loaded (by passing to Node app from browswer with websockets).
Redis 'game-events' messages now should include the token, so we know which
sockets to skip when pushing the update payload.